+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Unreleased
+
+* Fixed react template to add missing manifestPath to config.ts
+
 ## v1.0.0
 
 The official CLI for scaffolding, developing, and maintaining RapidREST projects. Available as both `rapidrest` and the shorter `rr` alias.